% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizing.R
\name{apply.size}
\alias{apply.size}
\title{Apply the local southern sizing method}
\usage{
apply.size(data, models, std.peaks)
}
\arguments{
\item{data}{List of data tables genereated by read.data}

\item{models}{List of nested models generated by get.models}

\item{std.peaks}{List of standard peak tables generated by get.std.peak}
}
\value{
The list of input data tables is returned with additional columns for model predictions,
   averages and finally the size in basepairs (column size.bps)
}
\description{
This function completes the local southern sizing method. See get.models for a description.
  First predictions for every data point for each of the 7 models are made.
  After that the appropriate rowwise averages of predictions are made using matricies.
  Finally the sizing is determined based on where each datapoint falls with respect to the standards.
}
\seealso{
\code{\link{get.models, precompute.avgs}}
}
