% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizing.R
\name{filter.size}
\alias{filter.size}
\title{Filter size}
\usage{
\method{filter}{size}(data)
}
\arguments{
\item{data}{List of data tables genereated by read.data and modified by apply.size.}
}
\value{
List of Filtered input data
}
\description{
This is a simple function to filter the sizes less than zero bp.
}
\seealso{
\code{\link{get.models, precompute.avgs}}
}
