% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{find.best.peaks}
\alias{find.best.peaks}
\title{Find the peaks in the standards channel}
\usage{
find.best.peaks(cand.peaks)
}
\arguments{
\item{cand.peaks}{table containing the candidate peaks where the number of peaks is more than 9.}
}
\value{
Dplyr table containing the best set of 9 peaks.
}
\description{
This function is used when more than 9 standard peaks are found.
   It fits all possible sets of linear models with 9 datapoints and choses the one that
   has the lowest deviance.
}
\seealso{
\code{\link{get.peaks,find.best.peak,read.data}}
}
