% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizing.R
\name{get.models}
\alias{get.models}
\title{Fit local southern sizing model}
\usage{
get.models(std.peaks)
}
\arguments{
\item{std.peaks}{List of standard peak tables generated by get.std.peak}
}
\value{
Returns a list of nested dataframes containing seven models for each data file. Each
   model is based on 3 datapoints. This will by used by apply.size function to finish the
   local southern sizing method.
}
\description{
Fits a series of 7 linear models necessary to use the local southern method
   of sizing, utilized by programs such as peak scanner. The local southern method sizes
   an unknown fragment by first computing 2 linear models. The first includes 2 standard points
   below the unknown fragment and one point above the unknown fragment. The second includes
   one point below the unknown fragment and two points above. Predictions from each model are
   then averaged to give the final size prediction.
}
\seealso{
\code{\link{get.std.peak, apply.size}}
}
