% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{get.peaks}
\alias{get.peaks}
\title{Find the peaks in the standards channel}
\usage{
get.peaks(ch5)
}
\arguments{
\item{ch5}{Column that contains the data from channel 5 (Std channel)}
}
\value{
Dplyr data tables containing the location and size of the
   standard peaks for each file.
}
\description{
This function locates candidate standard peaks using the findpeaks
   function in the pracma package. The inital peak thershold is set at
   1/6 of the maximum value in the standards channel. If 9 peaks are found at this
   thershold then the standards are returned. If less than 9 peaks are found then
   the thershold is iterativly lowered until at least 9 peaks are found. If more than
   9 peaks are found then models are fit for all possible sets of 9 peaks. The set of
   peaks with the highest R squared value is chosen as the true peaks. This function
   is called by get.std.peak.
}
\seealso{
\code{\link{get.std.peak,find.best.peak,read.data}}
}
