% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{get.std.peak}
\alias{get.std.peak}
\title{Find the peaks in the standards channel}
\usage{
get.std.peak(data.list)
}
\arguments{
\item{data.list}{List of data frames generated by the read.data function}
}
\value{
List of dplyr data tables containing the location and size of the
   standard peaks for each file.
}
\description{
This function identifies the standard peaks in the
   Liz channel (Channel 5). It uses the get peaks function to find the
   correct peaks.
}
\seealso{
\code{\link{get.peaks,find.best.peak,read.data}}
}
