% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizing.R
\name{precompute.avgs}
\alias{precompute.avgs}
\title{Rowwise averages of model predictions}
\usage{
precompute.avgs(data.subset)
}
\arguments{
\item{data.subset}{The data with model predictions added.}
}
\value{
The list of input data tables is returned with additional columns for model prediction averages.
}
\description{
This function is used by apply.size to get the averages needed for local southern sizing.
   It uses matrix algebra (the rowMeans function from the Matrix package), which is much much
   faster than doing a rowwise operation with dplyr.
}
\seealso{
\code{\link{get.models,apply.size}}
}
