% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read.abif2}
\alias{read.abif2}
\title{Read FSA file}
\usage{
read.abif2(filename, max.bytes.in.file = file.info(filename)$size,
  pied.de.pilote = 1.2, verbose = FALSE)
}
\arguments{
\item{filename}{The name of the file.}

\item{max.bytes.in.file}{The size in bytes of the file, defaulting to what is returned by file.info}

\item{pied.de.pilote}{Saftey factor: the argument readBin is set as pied.de.pilote*max.bytes.in.file.}

\item{verbose}{logical [FALSE]. If TRUE verbose mode is on.}
}
\value{
List of dplyr data tables containing the raw readings for
   each file in the directory
}
\description{
This function is a slightly modified version of read.abif from Fragman.
    Return is simplified to get rid of unnecessary meta data.
}
\references{
Covarrubias-Pazaran G, Diaz-Garcia L, Schlautman B, Salazar W, Zalapa J. Fragman: An R package for fragment analysis. 2016. BMC Genetics 17(62):1-8.
}
\seealso{
\code{\link{Fragman::read.abif}}
}
