% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read.data}
\alias{read.data}
\title{Read in all the FSA files in a directory}
\usage{
read.data(data.dir = getwd())
}
\arguments{
\item{data.dir}{input directory as character vector.
Defaults to the current working directory.}
}
\value{
List of dplyr data tables containing the raw readings for
   each file in the directory
}
\description{
This function reads in FSA data from the directory using a
   slightly modified version of the read_abif function in Fragman. It also
   does the pullup correction and smoothing using the pullup and transfft
   functions from Fragman.
}
\references{
Covarrubias-Pazaran G, Diaz-Garcia L, Schlautman B, Salazar W, Zalapa J. Fragman: An R package for fragment analysis. 2016. BMC Genetics 17(62):1-8.
}
\seealso{
\code{\link{read.abif2}}
}
